+++ libs/spark/src/Radio/Radio.spec.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { render, screen } from '@testing-library/react';
+import Radio from './Radio';
+
+describe('Radio', () => {
+  it('renders', () => {
+    render(<Radio inputProps={{ 'aria-label': 'label' }} />);
+
+    expect(screen.getByLabelText('label')).toBeTruthy();
+  });
+});