+++ libs/spark/src/Radio/RadioIcon.tsx
@@ -58,43 +58,27 @@ const RadioIconRoot = styled('span')(({ theme: { palette, transitions } }) => ({
   },
 }));
 
-// viewBox="0 0 26 26"
 const RadioIconCircle = createSvgIcon(
   <path d="M13 2C6.92487 2 2 6.92487 2 13C2 19.0751 6.92487 24 13 24C19.0751 24 24 19.0751 24 13C24 6.92487 19.0751 2 13 2ZM0 13C0 5.8203 5.8203 0 13 0C20.1797 0 26 5.8203 26 13C26 20.1797 20.1797 26 13 26C5.8203 26 0 20.1797 0 13Z" />,
-  'SparkRadioIconCircle'
+  'SparkRadioIconCircle',
+  '0 0 26 26'
 );
 
-// viewBox="0 0 26 26"
 const RadioIconDot = createSvgIcon(
   <path d="M13 21C17.4183 21 21 17.4183 21 13C21 8.58172 17.4183 5 13 5C8.58172 5 5 8.58172 5 13C5 17.4183 8.58172 21 13 21Z" />,
-  'SparkRadioIconDot'
+  'SparkRadioIconDot',
+  '0 0 26 26'
 );
 
-export default function RadioIcon({
-  checked,
-  fontSize,
-}: {
-  checked?: boolean;
-  fontSize?: 'small' | 'default';
-}) {
+export default function RadioIcon({ checked }: { checked?: boolean }) {
   return (
     <RadioIconRoot
       className={clsx('SparkRadioIcon-root', {
         'SparkRadioIcon-checked': checked,
       })}
     >
-      <RadioIconCircle
-        color="inherit"
-        fontSize={fontSize}
-        viewBox="0 0 26 26"
-        className="SparkRadioIcon-circle"
-      />
-      <RadioIconDot
-        color="inherit"
-        fontSize={fontSize}
-        viewBox="0 0 26 26"
-        className="SparkRadioIcon-dot"
-      />
+      <RadioIconCircle className="SparkRadioIcon-circle" />
+      <RadioIconDot className="SparkRadioIcon-dot" />
     </RadioIconRoot>
   );
 }