+++ libs/spark/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - **IconButton**
   - Can render without `ThemeProvider`.
+- **Radio**
+  - Can render without `ThemeProvider`.
 
 ## [v0.16.0](https://github.com/prenda-school/prenda-spark/compare/v0.15.0...v0.16.0) (2021-10-29)